Added Controls and more Functionality
Changed the mouse listener for the cell to be on the GameBoard.
Added methods to set the neighbors value of specific cells and the
adjacent cells.
Added method to insert a test shape.
Added functionality to change from painting single cells to shapes.
Added various shapes.

Added timer, start and stop buttons to run the simulation.
Added speed slider for selecting simulation speed.
Added button for toggling between painting cells and shapes.
Added shape selector for painting.
Added generation counter for displaying the iteration of simulation it
is in. 

Overall cleaned some things up to be more efficient and better looking
(code and GUI wise);